requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
streamlit>=1.28.0

# Database
sqlalchemy>=2.0.0
psycopg[binary,pool]>=3.1.0
alembic>=1.12.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0

# AI and ML
openai>=1.3.0
scikit-learn>=1.3.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# HTTP and API
requests>=2.31.0
httpx>=0.25.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Optional: For enhanced features
# ollama>=0.1.0  # For local AI models
# redis>=5.0.0   # For caching
# celery>=5.3.0  # For background tasks
